+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 * feature#97: Add FreeBSD service control
 * feature#111: Add a Template Export Button to make sharing Template information with the Cacti Group Easy
 * feature#112: Provide a means to view the DNS Cache and Prune/Edit Entries
+* feature#113: Provide buttons 'New', 'Rename', 'Delete' from the FlowView Tab
 * feature#114: Provide an External Address dropdown in the Filter Definition
 * feature#116: Have flow-capture automatically restart collectors that have crashed
 * feature#117: Allow guest account access